+++ recipes/proxy.rb
@@ -65,3 +65,13 @@ if install_via == 'systemd'
     action [:create, :enable, :start]
   end
 end
+
+firewall_rule 'kube_proxy' do
+  port node['kubernetes']['proxy']['global']['metrics_port']
+  protocol :tcp
+  interface node['kubernetes']['interface']
+  command :allow
+  only_if do
+    node['kubernetes']['enable_firewall']
+  end
+end